requirements.txt
# Dev environment
pip
autopep8

# App
discord.py
python-dotenv
